+++ src/SW_Output_outtext.c
@@ -186,18 +186,18 @@ static void get_outstrheader(OutPeriod pd, char *str, size_t sizeof_str) {
 */
 static void _create_filename_ST(char *str, char *flag, int iteration,
 			char *filename, size_t sizeof_filename, LOG_INFO* LogInfo) {
+    int startIndex = 0, strLen = 0; // For `sw_strtok()`
+
 	char *basename;
 	char *ext;
-	char *fileDup = (char *)Mem_Malloc(strlen(str) + 1,
-							"_create_filename_ST", LogInfo);
+	char *fileDup = Str_Dup(str, LogInfo);
     if(LogInfo->stopRun) {
         return; // Exit function prematurely due to error
     }
 
 	// Determine basename and file extension
-	strcpy(fileDup, str); // copy file name to new variable
-	basename = strtok(fileDup, ".");
-	ext = strtok(NULL, ".");
+	basename = sw_strtok(fileDup, &startIndex, &strLen, ".");
+	ext = sw_strtok(fileDup, &startIndex, &strLen, ".");
 
 	// Put new file together
 	if (iteration > 0) {