+++ NEWS.md
@@ -1,6 +1,6 @@
 # NEWS
 
-# SOILWAT2 v7.2.0-devel
+# SOILWAT2 v7.2.0
 * Simulation output remains the same as the previous version.
 
 * SOILWAT2 now handles errors more gracefully (#346; @N1ckP3rsl3y, @dschlaep).
@@ -13,6 +13,9 @@
   This allows to hide implementation details, i.e., STEPWAT2 will no
   longer need to interact directly with `pcg` and rSOILWAT2 will no longer
   depend on `pcg` which it never has used.
+
+* New `sw_strtok()` is thread-safe and replaces not thread-safe `strtok()`
+  (#376; @N1ckP3rsl3y).
 
 * Obsolete code in DEBUG_MEM* sections is removed (#369; @dschlaep).